# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Open Payments Split API",
    "version": "0.1.0",
    "description": "Endpoints de split checkout (single/group) y callback"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
  "paths": {
    "/api/split/checkout": {
      "post": {
        "summary": "Init split checkout (merchant + plataforma)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerAddress": { "type": "string" },
                  "merchantAddress": { "type": "string" },
                  "amountMinor": { "type": "integer" },
                  "split": {
                    "type": "object",
                    "properties": {
                      "merchantPct": { "type": "integer" },
                      "platformPct": { "type": "integer" }
                    }
                  }
                },
                "required": ["customerAddress", "merchantAddress", "amountMinor"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL de redirección al IdP y nonce",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "redirectUrl": { "type": "string" },
                    "nonce": { "type": "string" }
                  },
                  "required": ["redirectUrl", "nonce"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/op/callback": {
      "get": {
        "summary": "Callback tras consentimiento para crear outgoing payments",
        "parameters": [
          { "name": "interact_ref", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "nonce", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Pagos creados",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CallbackResult" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/split/group-checkout": {
      "post": {
        "summary": "Init group checkout (varios payers)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "merchantAddress": { "type": "string" },
                  "totalAmountMinor": { "type": "integer" },
                  "payers": { "type": "array", "items": { "type": "string" } }
                },
                "required": ["merchantAddress", "totalAmountMinor", "payers"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Listado de redirecciones por payer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "merchant": { "type": "string" },
                    "totalMinor": { "type": "integer" },
                    "count": { "type": "integer" },
                    "results": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/GroupInitResult" }
                    }
                  },
                  "required": ["merchant", "totalMinor", "count", "results"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GroupInitResult": {
        "type": "object",
        "properties": {
          "payer": { "type": "string" },
          "shareMinor": { "type": "integer" },
          "redirectUrl": { "type": "string" },
          "nonce": { "type": "string" }
        },
        "required": ["payer", "shareMinor", "redirectUrl", "nonce"]
      },
      "CallbackResult": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "payer": { "type": "string" },
          "outgoingPayments": { "type": "array", "items": { "type": "object" } }
        },
        "required": ["status", "outgoingPayments"]
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Solicitud inválida",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  }
}